+++ src/main/java/com/jarqprog/artGallery/rest/controller/SimpleContactController.java
@@ -1,4 +1,4 @@
-package com.jarqprog.artGallery.rest.restController;
+package com.jarqprog.artGallery.rest.controller;
 
 import com.jarqprog.artGallery.domain.dto.ContactDTO;
 import com.jarqprog.artGallery.domain.useCases.ContactService;
