+++ src/main/java/com/jarqprog/artGallery/springWebMVC/controller/RegistrationController.java
@@ -30,7 +30,7 @@ public class RegistrationController {
     public String showRegistrationForm(WebRequest request, Model model) {
         RegistrationDTO registrationDTO = new RegistrationDTO();
         model.addAttribute("registration", registrationDTO);
-        return "/user/registration";
+        return "user/registration";
     }
 
     @PostMapping
